content/pyproject.toml
[tool.poetry]
name = "content api"
version = "0.1.0"
description = "api of lms's content"
authors = ["Nozir Tukhramurodov <nozir9407@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
asyncpg = "^0.29.0"
fastapi = "^0.113.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
pydantic-core = "^2.22.0"
python-dotenv = "^1.0.0"
SQLAlchemy = "^2.0.31"
uvicorn = "^0.30.5"
async-fastapi-jwt-auth = "^0.6.6"
gunicorn = "^22.0.0"
redis = "^5.0.8"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
fastapi-pagination = "^0.12.26"
aiohttp = "^3.10.5"
python-slugify = "^8.0.4"
phonenumbers = "^8.13.42"
orjson= "^3.9.10"
faker = "^22.5.1"
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
flake8-bugbear = "^24.1.17"
flake8-isort = "^6.1.1"
flake8-bandit = "^4.1.1"
flake8-pytest-style = "^1.7.2"
flake8-pytest = "^1.4"
flake8-print = "^5.0.0"
flake8-commas = "^2.1.0"
flake8-broken-line = "^1.0.0"
isort = "^5.13.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
faker = "^22.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
